fix: SchedulerRef schedule

+++ mobile/src/main/java/fit/asta/health/scheduler/data/db/entity/AlarmSync.kt
@@ -8,7 +8,7 @@ import androidx.room.PrimaryKey
 data class AlarmSync(
     @PrimaryKey
     @ColumnInfo(name = "id")
-    var alarmId: Int=0,
+    var alarmId: Long = 0,
     @ColumnInfo(name = "scheduleId")
-    var scheduleId: String=""
+    var scheduleId: String = ""
 )
